deleting module
