--- src/crumbs/raster_utils.py
@@ -1,18 +1,0 @@
-#!/usr/bin/python
-
-def random_sample_from_masked_array(masked, nb_sample):
-    """ Sample indices uniformely at random in a masked array, ignoring masked values.
-        Returns a tuple (idx,idy)
-    """
-    import numpy as np
-     #Assign False = 0, True = 1
-    weights =~ masked.mask + 0
-    normalized = weights.ravel()/float(weights.sum())
-    index = np.random.choice(
-        masked.size,
-        size=nb_sample,
-        replace=False,
-        p=normalized
-    )
-    idy, idx = np.unravel_index(index, masked.shape)
-    return idx, idy
