+++ src/crumbs/sample.py
@@ -1,5 +1,22 @@
 #!/usr/bin/python
 
+def random_sample_from_masked_array(masked, nb_sample):
+    """ Sample indices uniformely at random in a masked array, ignoring masked values.
+        Returns a tuple (idx,idy)
+    """
+    import numpy as np
+     #Assign False = 0, True = 1
+    weights =~ masked.mask + 0
+    normalized = weights.ravel()/float(weights.sum())
+    index = np.random.choice(
+        masked.size,
+        size=nb_sample,
+        replace=False,
+        p=normalized
+    )
+    idy, idx = np.unravel_index(index, masked.shape)
+    return idx, idy
+
 def uniform_integer(min, max):
     import random
     print(random.randint(int(min), int(max)))
@@ -9,7 +26,6 @@ def uniform_real(min, max):
     print(random.uniform(float(min), float(max)))
 
 def uniform_latlon(raster_path, band):
-    from raster_utils import random_sample_from_masked_array
     from os.path import exists
     import rasterio
     assert exists(raster_path), 'File doest not exists:' + raster_path
