+++ src/crumbs/sdm.py
@@ -1,5 +1,22 @@
 #!/usr/bin/python
 
+def random_sample_from_masked_array(masked, nb_sample):
+    """ Sample indices uniformely at random in a masked array, ignoring masked values.
+        Returns a tuple (idx,idy)
+    """
+    import numpy as np
+     #Assign False = 0, True = 1
+    weights =~ masked.mask + 0
+    normalized = weights.ravel()/float(weights.sum())
+    index = np.random.choice(
+        masked.size,
+        size=nb_sample,
+        replace=False,
+        p=normalized
+    )
+    idy, idx = np.unravel_index(index, masked.shape)
+    return idx, idy
+
 def present_geopanda(gdf):
     print("        - number of duplicates: ", gdf.duplicated(subset='geometry', keep='first').sum())
     print("        - number of NA's: ", gdf['geometry'].isna().sum())
@@ -23,7 +40,6 @@ def sample_background(demRaster, nb_sample=30):
     """
     import rasterio
     import geopandas
-    from raster_utils import random_sample_from_masked_array
     with rasterio.open(demRaster) as src:
         Z = src.read(1, masked=True)
         cols, rows = random_sample_from_masked_array(Z, nb_sample)
